+++ src/lightning/app/__init__.py
@@ -27,15 +27,15 @@ if os.path.isfile(os.path.join(os.path.dirname(__file__), "__about__.py")):
     from lightning.app.__about__ import *  # noqa: F401, F403
 if "__version__" not in locals():
     if os.path.isfile(os.path.join(os.path.dirname(__file__), "__version__.py")):
-        from lightning.app.__version__ import version as __version__
+        pass
     elif package_available("lightning"):
         from lightning import __version__  # noqa: F401
 
 from lightning.app.core.app import LightningApp  # noqa: E402
 from lightning.app.core.flow import LightningFlow  # noqa: E402
-from lightning.app.core.plugin import LightningPlugin  # noqa: E402
 from lightning.app.core.work import LightningWork  # noqa: E402
 from lightning.app.perf import pdb  # noqa: E402
+from lightning.app.plugin.plugin import LightningPlugin  # noqa: E402
 from lightning.app.utilities.packaging.build_config import BuildConfig  # noqa: E402
 from lightning.app.utilities.packaging.cloud_compute import CloudCompute  # noqa: E402
