+++ src/lightning/app/core/__init__.py
@@ -1,6 +1,5 @@
 from lightning.app.core.app import LightningApp
 from lightning.app.core.flow import LightningFlow
-from lightning.app.core.plugin import LightningPlugin
 from lightning.app.core.work import LightningWork
 
-__all__ = ["LightningApp", "LightningFlow", "LightningWork", "LightningPlugin"]
+__all__ = ["LightningApp", "LightningFlow", "LightningWork"]
