+++ src/lightning/app/plugin/__init__.py
@@ -1,0 +1,4 @@
+import lightning.app.plugin.actions as actions
+from lightning.app.plugin.plugin import LightningPlugin
+
+__all__ = ["LightningPlugin", "actions"]
