+++ tests/tests_app/plugin/test_plugin.py
@@ -9,7 +9,7 @@ import pytest
 from fastapi import status
 from fastapi.testclient import TestClient
 
-from lightning.app.core.plugin import _Run, _start_plugin_server
+from lightning.app.plugin.plugin import _Run, _start_plugin_server
 
 
 @pytest.fixture()
